Update default tracker with ts & listeners
